Add files via upload
# creating_knowledge_graph_NLP
The Knowledge Graph Construction is a Python project that aims to extract valuable information and relationships from a collection of a website.
# Key Features
Extraction of Headlines: The project includes functionality to extract relevant headlines from URLs. It employs techniques such as parsing and text cleaning to obtain clean and informative headlines.

Named Entity Recognition: The project utilizes the power of the spaCy library for named entity recognition. It identifies and labels entities such as organizations, locations, persons, and other relevant entities present in the headlines.

Relationship Extraction: Using rule-based parsing and dependency analysis, the project extracts the relationships between the identified entities in the headlines. It identifies the main verb in each sentence as the predicate and establishes the grammatical relationships between entities.

Knowledge Graph Visualization: The extracted entities and relationships are used to construct a knowledge graph. This graph provides a visual representation of the interconnected entities and their relationships, enabling a better understanding of the underlying information.

Graph Exploration: The project allows for interactive exploration of the constructed knowledge graph. Users can specify a relationship of interest and visualize the corresponding subgraph, providing insights into specific aspects of the data.